+++ gulpfile.js
@@ -9,7 +9,6 @@ var gulp         = require('gulp'),
     concat       = require('gulp-concat'),
     server       = require('gulp-webserver'),
     runsSequence = require('run-sequence'),
-    traceur         = require('gulp-traceur'),
     gutil        = require('gulp-util');
 
 
@@ -43,14 +42,8 @@ gulp.task('test', function() {
         .pipe(gulp.dest('./build/js'))
 });
 
-gulp.task('traceur:runtime', function() {
-    return gulp.src(traceur.RUNTIME_PATH)
-        .pipe(gulp.dest('./src/js/core/traceur'));
-});
-
-
 // copy local index file to build folder
-gulp.task('vendor', ['traceur:runtime'], function() {
+gulp.task('vendor', [], function() {
     return gulp.src(['./src/js/core/**'])
         .pipe(gulp.dest('./build/js/vendor'))
         .pipe(gulp.dest('./lib/vendor'));